+++ app/src/main/resources/ApplicationResources.properties
@@ -429,13 +429,9 @@ frontpageConfig.weblogs.error=Unexpected error accessing Weblogs
 
 # --------------------------------------------------------------- Invite member
 
-inviteMember.title=Invite New Member
-
 inviteMember.subtitle=Invite a new user to join this weblog
-inviteMember.prompt=Invite a user to join this weblog by selecting his or \
-her email address below, picking a permission for them and clicking the send \
-invitation button. If the user is not shown in the list below, type the first \
-couple of letters of the username or email address view matches.
+inviteMember.prompt=Invite a user to join this weblog by selecting the user,\
+  picking a permission for the new user and clicking the send invitation button.
 
 inviteMember.userName=Username
 inviteMember.administrator=Owner
@@ -719,9 +715,6 @@ memberPermissions.membersRemoved=Removed {0} user(s)
 memberPermissions.membersChanged=Changed permission for {0} user(s)
 
 memberPermissions.saveError=Error saving permissions
-
-memberPermissions.inviteMember=Invite new member
-memberPermissions.whyInvite=Invite somebody to help you write your weblog.
 
 memberPermissions.permissionsHelpTitle=Permission Definitions
 memberPermissions.permissionHelp=\
