+++ app/src/main/resources/ApplicationResources_de.properties
@@ -329,9 +329,8 @@ inviteMember.limited=Begrenzt
 inviteMember.notificationContent=Sie wurden eingeladen dem Weblog "{0}" ({1}) beizutreten. Klicken Sie auf den folgenden Link und loggen Sie sich als Benutzer [{2}] ein oder lehnen Sie diese Einladung ab <{3}>.
 inviteMember.notificationSubject=TightBlog: Einladung zu Weblog "{0}" ({1})
 inviteMember.permissions=Berechtigungen
-inviteMember.prompt=Laden Sie einen neuen Benutzer zu diesem Weblog ein indem Sie unten seine E-Mail-Adresse und eine Berechtigung ausw\u00E4hlen und anschlie\u00DFend auf den Absenden-Button klicken. Wenn der Benutzer nicht in der Liste unten zu sehen ist, tippen Sie bitte die ersten paar Buchstaben des Benutzernamens oder der E-Mail-Adresse ein.
+inviteMember.prompt=Laden Sie einen neuen Benutzer zu diesem Weblog ein indem Sie unten seine E-Mail-Adresse und eine Berechtigung ausw\u00E4hlen und anschlie\u00DFend auf den Absenden-Button klicken.
 inviteMember.subtitle=Laden Sie einen neuen Benutzer zu diesem Weblog ein
-inviteMember.title=Ein Mitglied einladen
 inviteMember.userInvited=Der Benutzer wurde erfolgreich eingeladen
 inviteMember.userName=Benutzername
 # ------------------------------------------------------ Invitations management
@@ -398,7 +397,6 @@ maintenance.title=Webseitenwartung
 memberPermissions.administrator=Admin
 memberPermissions.author=Autor
 memberPermissions.description=Sie k\u00F6nnen die Rechte der Weblog-Mitglieder \u00E4ndern und oder die Benutzer l\u00F6schen indem Sie die Verwaltungsoberfl\u00E4che in der unten folgenden Tabelle verwenden und abschlie\u00DFend Speichern anklicken um die \u00C4nderungen umzusetzen. Bedenken Sie, dass Sie sich nicht selbst degradieren oder l\u00F6schen k\u00F6nnen.
-memberPermissions.inviteMember=Mitglieder einladen
 memberPermissions.limited=Beschr\u00E4nkt
 memberPermissions.membersChanged=Rechte ge\u00E4ndert f\u00FCr {0} Benutzer.
 memberPermissions.membersRemoved={0} Benutzer entfernt
@@ -408,7 +406,6 @@ memberPermissions.remove=Entfernen
 memberPermissions.subtitle=Rechte verwalten von Benutzern in Weblog <span>{0}</span>
 memberPermissions.title=Weblog Benutzerberechtigungen
 memberPermissions.userName=Benutzername
-memberPermissions.whyInvite=Laden Sie jemanden ein, der Ihnen beim Schreiben des Weblogs helfen soll.
 # ----------------------------------------------------------- Member permissions
 # ------------------------------------------------------------ Navigation Bar
 navigationBar.login=Einloggen
