+++ app/src/main/resources/ApplicationResources_es.properties
@@ -172,9 +172,8 @@ error.permissionDenied.reason2=Ha entrado usando las may\u00FAsculas/min\u00FAsc
 error.permissionDenied.reason3=Su conexi\u00F3n a la base de datos del servidor est\u00E1 mal configurada. Para resolver este problema, consulte al administrador del sistema.
 errorPage.title=Excepci\u00F3n inesperada
 errorPage.message=TightBlog ha encontrado y registrado una excepci\u00F3n inesperada.
-inviteMember.title=Invitar a un nuevo miembro
 inviteMember.subtitle=Invitar a un nuevo usuario a que se una a este weblog
-inviteMember.prompt=Invitar a un usuario a unirse a este weblog seleccionando abajo su direcci\u00F3n de correo electr\u00F3nico, eligiendo un permiso para este y haciendo click en el bot\u00F3n Enviar invitaci\u00F3n. Si el usuario no se muestra en la lista de abajo, escriba las primeras letras del nombre de usuario o direcci\u00F3n de correo electr\u00F3nico para encontrar coincidencias.
+inviteMember.prompt=Invitar a un usuario a unirse a este weblog seleccionando abajo su direcci\u00F3n de correo electr\u00F3nico, eligiendo un permiso para este y haciendo click en el bot\u00F3n Enviar invitaci\u00F3n.
 inviteMember.userName=Nombre de usuario
 inviteMember.administrator=Administrador
 inviteMember.author=Autor
@@ -240,10 +239,8 @@ memberPermissions.administrator=Administrador
 memberPermissions.author=Autor
 memberPermissions.limited=Limitado
 memberPermissions.remove=Eliminar
-memberPermissions.inviteMember=Invitar a un nuevo miembro
 memberPermissions.membersRemoved={0} usuario(s) eliminado(s)
 memberPermissions.membersChanged=Permiso cambiado para {0} usuario(s)
-memberPermissions.whyInvite=Invitar a alguien para ayudar a escribir el weblog.
 memberPermissions.permissionsHelpTitle=\u00BFQu\u00E9 son los permisos?
 memberPermissions.permissionHelp=Los permisos de <b>administrador</b> permiten a los usuarios enviar entradas, administrar el weblog y sus usuarios<br />Los permisos de <b>autor</b> permiten a los usuarios enviar entradas, pero no administrar el weblog<br />Los permisos <b>limitados</b> permiten a los usuarios guardar entradas solamente de tipo borrador.
 newUser.addNewUser=Registro de nuevo usuario
