+++ app/src/main/resources/ApplicationResources_fr.properties
@@ -366,12 +366,10 @@ Quand vous aurez terminé, vous ne verrez enfin plus l'agaçante page actuelle.
 
 # --------------------------------------------------------------- Invite member
 
-inviteMember.title=Inviter un nouveau membre
 inviteMember.subtitle=Inviter un nouvel utilisateur à joindre ce blog
 inviteMember.prompt=Invitez un utilisateur à rejoindre ce blog en sélectionnant \
 son adresse de courriel ci-dessous et en choisissant une des permissions d'accés \
-pour celui-ci. Cliquez ensuite sur le bouton 'Inviter'. Si l'utilisateur n'est pas visible \
-dans la liste, entrez les premières lettres de son nom ou de son courriel afin de le retrouver.
+pour celui-ci. Cliquez ensuite sur le bouton 'Inviter'.
 
 inviteMember.userName=Nom d'utilisateur
 inviteMember.administrator=Administrateur
@@ -480,9 +478,6 @@ memberPermissions.remove=Effacer
 
 memberPermissions.membersRemoved={0} utilisateurs enlevés.
 memberPermissions.membersChanged=Les permissions d'accés de {0} utilisateur(s) ont été changées.
-
-memberPermissions.inviteMember=Inviter un nouveau membre
-memberPermissions.whyInvite=Inviter quelqu'un à joindre votre blog et vous aider à écrire.
 
 memberPermissions.permissionsHelpTitle=Qu'est-ce qu'une permission ou droit d'accés ?
 memberPermissions.permissionHelp=\
