+++ app/src/main/resources/ApplicationResources_ja.properties
@@ -297,10 +297,7 @@ TightBlog\u3067\u306F\u7279\u5B9A\u306E\u30D6\u30ED\u30B0\u3092\u30B5\u30A4\u30C
 # --------------------------------------------------------------- Invite member
 
 inviteMember.subtitle=\u3053\u306E\u30D6\u30ED\u30B0\u3078\u4ED6\u306E\u30E6\u30FC\u30B6\u3092\u62DB\u5F85\u3057\u307E\u3059
-inviteMember.prompt=\u30B7\u30B9\u30C6\u30E0\u306B\u767B\u9332\u3057\u3066\u3044\u308B\u4ED6\u306E\u30E6\u30FC\u30B6\u3092\u3053\u306E\u30D6\u30ED\u30B0\u306E\u30E1\u30F3\u30D0\u3068\u3059\u308B\u3053\u3068\u304C\u3067\u304D\u307E\u3059\u3002\
-\u30E6\u30FC\u30B6\u540D\u3068\u3042\u305F\u3048\u308B\u6A29\u9650\u3092\u9078\u629E\u3057\u3066\u300C\u62DB\u5F85\u72B6\u3092\u767A\u9001\u3059\u308B\u300D\u3092\u30AF\u30EA\u30C3\u30AF\u3057\u3066\u304F\u3060\u3055\u3044\u3002\
-\u300C\u30E6\u30FC\u30B6\u540D\u300D\u6B04\u3067\u306F\u30E6\u30FC\u30B6\u540D\u306A\u3044\u3057\u30E1\u30FC\u30EB\u30A2\u30C9\u30EC\u30B9\u306E\u6700\u521D\u306E\u3064\u3065\u308A\u3092\u30BF\u30A4\u30D7\u3059\u308B\u3053\u3068\u3067\u3001\
-\u3042\u306A\u305F\u304C\u62DB\u5F85\u3057\u305F\u3044\u30E6\u30FC\u30B6\u306E\u7D5E\u308A\u3053\u307F\u304C\u3067\u304D\u307E\u3059\u3002
+inviteMember.prompt=\u30b7\u30b9\u30c6\u30e0\u306b\u767b\u9332\u3057\u3066\u3044\u308b\u4ed6\u306e\u30e6\u30fc\u30b6\u3092\u3053\u306e\u30d6\u30ed\u30b0\u306e\u30e1\u30f3\u30d0\u3068\u3059\u308b\u3053\u3068\u304c\u3067\u304d\u307e\u3059\u3002\u30e6\u30fc\u30b6\u540d\u3068\u3042\u305f\u3048\u308b\u6a29\u9650\u3092\u9078\u629e\u3057\u3066\u300c\u62db\u5f85\u72b6\u3092\u767a\u9001\u3059\u308b\u300d\u3092\u30af\u30ea\u30c3\u30af\u3057\u3066\u304f\u3060\u3055\u3044\u3002
 
 inviteMember.userName=\u30E6\u30FC\u30B6\u540D
 inviteMember.administrator=\u7BA1\u7406\u8005
@@ -400,9 +397,6 @@ memberPermissions.remove=\u524A\u9664
 
 memberPermissions.membersRemoved={0}\u4EBA\u306E\u30E6\u30FC\u30B6\u3092\u524A\u9664\u3057\u307E\u3057\u305F
 memberPermissions.membersChanged={0}\u4EBA\u306E\u30E6\u30FC\u30B6\u6A29\u9650\u3092\u5909\u66F4\u3057\u307E\u3057\u305F
-
-memberPermissions.inviteMember=\u65B0\u3057\u3044\u30E6\u30FC\u30B6\u3092\u62DB\u5F85\u3059\u308B
-memberPermissions.whyInvite=\u30D6\u30ED\u30B0\u57F7\u7B46\u3092\u624B\u52A9\u3051\u3057\u3066\u304F\u308C\u308B\u4EBA\u3092\u62DB\u5F85\u3057\u307E\u3059
 
 memberPermissions.permissionsHelpTitle=\u6A29\u9650\u306B\u3064\u3044\u3066
 memberPermissions.permissionHelp=\
@@ -919,7 +913,6 @@ planetSubscriptions.title=\u30B5\u30D6\u30B9\u30AF\u30EA\u30D7\u30B7\u30E7\u30F3
 mediaFile.delete.confirm=\u9078\u629E\u3055\u308C\u305F\u30D5\u30A1\u30A4\u30EB\u3092\u524A\u9664\u3057\u3066\u3088\u308D\u3057\u3044\u3067\u3059\u304B?
 commentManagement.saveError=\u30B3\u30E1\u30F3\u30C8\u306E\u4FDD\u5B58\u306B\u5931\u6557\u3057\u307E\u3057\u305F
 comments.ldapAuthenticatorUserName=\u3042\u306A\u305F\u306EActive Directory\u30E6\u30FC\u30B6\u540D\u3092\u5165\u529B\u3057\u3066\u304F\u3060\u3055\u3044\u3002
-inviteMember.title=\u65B0\u3057\u3044\u30E1\u30F3\u30D0\u30FC\u306E\u62DB\u5F85
 websiteRemove.error=\u30D6\u30ED\u30B0 [{0}] \u306E\u524A\u9664\u30A8\u30E9\u30FC
 weblogEdit.mediaCastFailedFetchingInfo=\u30A8\u30F3\u30AF\u30ED\u30FC\u30B8\u30E3\u306B\u30A2\u30AF\u30BB\u30B9\u3067\u304D\u307E\u305B\u3093\u3067\u3057\u305F\u3002URL\u5185\u306E\u30DB\u30B9\u30C8\u540D\u3092\u30C1\u30A7\u30C3\u30AF\u3057\u3066\u304F\u3060\u3055\u3044\u3002
 templates.error.actionNull=\u30A2\u30AF\u30B7\u30E7\u30F3\u306F\u5FC5\u9808\u9805\u76EE\u3067\u3059
