+++ app/src/main/resources/ApplicationResources_ko.properties
@@ -354,13 +354,9 @@ index.setFrontpageHelp=\
 
 # --------------------------------------------------------------- Invite member
 
-inviteMember.title=\uc0c8 \uba64\ubc84 \ucd08\ub300
-
 inviteMember.subtitle=\uc0c8 \uc0ac\uc6a9\uc790\ub97c \uc774 \uc6f9\ub85c\uadf8\ub85c \ucd08\ub300
 inviteMember.prompt=\uc544\ub798 \uc804\uc790\uc6b0\ud3b8 \uc8fc\uc18c\ub97c \uc120\ud0dd\ud558\uc5ec, \uc0c8 \uc0ac\uc6a9\uc790\ub97c \uc774 \uc6f9\ub85c\uadf8\ub85c \ucd08\ub300\ud558\uc2ed\uc2dc\uc624. \
-\uc0c8 \uc0ac\uc6a9\uc790\ub4e4\uc744 \uc704\ud55c \uad8c\ud55c\uc744 \uace0\ub974\uc2dc\uace0, \ucd08\ub300 \ubc1c\uc1a1 \ubc84\ud2bc\uc744 \ud074\ub9ad\ud558\uc2ed\uc2dc\uc624. \
-\uc0ac\uc6a9\uc790\uac00 \uc544\ub798 \ubaa9\ub85d\uc5d0 \ubcf4\uc774\uc9c0 \uc54a\uc73c\uba74, \uc0ac\uc6a9\uc790 \uc774\ub984\uc774\ub098 \uc804\uc790\uc6b0\ud3b8 \uc8fc\uc18c\uc758 \
-\uc55e \ub450 \uae00\uc790\ub97c \uc785\ub825\ud558\uc2ed\uc2dc\uc624.
+\uc0c8 \uc0ac\uc6a9\uc790\ub4e4\uc744 \uc704\ud55c \uad8c\ud55c\uc744 \uace0\ub974\uc2dc\uace0, \ucd08\ub300 \ubc1c\uc1a1 \ubc84\ud2bc\uc744 \ud074\ub9ad\ud558\uc2ed\uc2dc\uc624.
 
 inviteMember.userName=\uc0ac\uc6a9\uc790\uba85
 inviteMember.administrator=\uad00\ub9ac\uc790
@@ -529,9 +525,6 @@ memberPermissions.remove=\uc0ad\uc81c
 
 memberPermissions.membersRemoved={0} \uba85\uc758 \uc0ac\uc6a9\uc790 \uc81c\uac70
 memberPermissions.membersChanged={0} \uba85\uc758 \uc0ac\uc6a9\uc790 \uad8c\ud55c \ubcc0\uacbd
-
-memberPermissions.inviteMember=\uc0c8 \uba64\ubc84 \ucd08\ub300
-memberPermissions.whyInvite=\uc6f9\ub85c\uadf8 \uc791\uc131\uc744 \ub3d5\uae30 \uc704\ud574 \ub2e4\ub978 \uc0ac\ub78c \ucd08\ub300
 
 memberPermissions.permissionsHelpTitle=\uad8c\ud55c\uc774\ub780?
 memberPermissions.permissionHelp=\
