+++ app/src/main/resources/ApplicationResources_ru.properties
@@ -224,7 +224,6 @@ errorPage.message=TightBlog \u0432\u0441\u0442\u0440\u0435\u0442\u0438\u043B \u0
 # ----------------------------------------------------------------- Invite member
 
 inviteMember.subtitle=\u041F\u0440\u0438\u0433\u043B\u0430\u0448\u0435\u043D\u0438\u0435 \u043D\u043E\u0432\u043E\u0433\u043E \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F \u043F\u0440\u0438\u0441\u043E\u0435\u0434\u0438\u043D\u0438\u0442\u044C\u0441\u044F \u043A \u044D\u0442\u043E\u043C\u0443 \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0443
-inviteMember.prompt= \u041F\u0440\u0438\u0433\u043B\u0430\u0448\u0435\u043D\u0438\u0435 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F \u043F\u0440\u0438\u0441\u043E\u0435\u0434\u0438\u043D\u0438\u0442\u044C\u0441\u044F \u043A \u044D\u0442\u043E\u043C\u0443 \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0443, \u0432\u044B\u0431\u0440\u0430\u0432 \u0435\u0433\u043E \ \u0438\u043B\u0438 \u043F\u043E email \u0430\u0434\u0440\u0435\u0441\u0443, \u0443\u043A\u0430\u0437\u0430\u043D\u043D\u043E\u043C\u0443 \u043D\u0438\u0436\u0435, \u0434\u0430\u0432 \u0440\u0430\u0437\u0440\u0435\u0448\u0435\u043D\u0438\u0435 \u0434\u043B\u044F \u043D\u0438\u0445 \u0438 \u0438 \u043A\u043B\u0438\u043A\u043D\u0443\u0432  \u043A\u043D\u043E\u043F\u043A\u0443 "\u041E\u043F\u0440\u0430\u0432\u0438\u0442\u044C"  \ "\u041F\u0440\u0438\u0433\u043B\u0430\u0441\u0438\u0442\u044C".  \u0415\u0441\u043B\u0438 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C \u043D\u0435 \u043E\u0442\u043E\u0431\u0440\u0430\u0436\u0430\u0435\u0442\u0441\u044F \u0432 \u043D\u0438\u0436\u0435\u043F\u0435\u0440\u0435\u0447\u0438\u0441\u043B\u0435\u043D\u043D\u043E\u043C \u0441\u043F\u0438\u0441\u043A\u0435, \u043D\u0430\u0431\u0435\u0440\u0438\u0442\u0435 \u043F\u0435\u0440\u0432\u044B\u0435 \u043D\u0435\u0441\u043A\u043E\u043B\u044C\u043A\u043E \u0431\u0443\u043A\u0432  \u0438\u043C\u0435\u043D\u0438 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F \u0438\u043B\u0438 \u0435\u0433\u043E email \u0430\u0434\u0440\u0435\u0441, \u0447\u0442\u043E\u0431\u044B \u0443\u0432\u0438\u0434\u0435\u0442\u044C \u0441\u043E\u0432\u043F\u0430\u0434\u0435\u043D\u0438\u044F.
 inviteMember.userName=\u0438\u043C\u044F \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F
 inviteMember.administrator=\u0410\u0434\u043C\u0438\u043D\u0438\u0441\u0442\u0440\u0430\u0442\u043E\u0440
 inviteMember.author=\u0410\u0432\u0442\u043E\u0440
@@ -317,12 +316,8 @@ memberPermissions.administrator=\u0410\u0434\u043C\u0438\u043D\u0438\u0441\u0442
 memberPermissions.author=\u0410\u0432\u0442\u043E\u0440
 memberPermissions.limited=\u041E\u0433\u0440\u0430\u043D\u0438\u0447\u0435\u043D (\u041E\u0433\u0440\u0430\u043D\u0438\u0447\u0435\u043D\u043D\u044B\u0439)
 memberPermissions.remove=\u0423\u0434\u0430\u043B\u0438\u0442\u044C
-memberPermissions.inviteMember=\u041F\u0440\u0438\u0433\u043B\u0430\u0441\u0438\u0442\u044C \u043D\u043E\u0432\u043E\u0433\u043E \u0447\u043B\u0435\u043D\u0430
-
 memberPermissions.membersRemoved=\u0423\u0434\u0430\u043B\u0435\u043D\u043D\u044B\u0439 (-\u044B\u0435) {0} \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C (-\u043B\u0438)
 memberPermissions.membersChanged=\u0418\u0437\u043C\u0435\u043D\u0435\u043D\u043D\u044B\u0435 \u043F\u0440\u0430\u0432\u0430 \u0434\u043B\u044F {0}\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F (-\u043B\u0435\u0439)
-memberPermissions.whyInvite=\u041F\u0440\u0438\u0433\u043B\u0430\u0441\u0438\u0442\u044C "\u043F\u043E\u043C\u043E\u0449\u043D\u0438\u043A\u0430" \u0434\u043B\u044F \u0432\u0435\u0434\u0435\u043D\u0438\u044F \u0441\u0432\u043E\u0435\u0433\u043E \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0430.
-
 memberPermissions.permissionsHelpTitle=\u0427\u0442\u043E \u0442\u0430\u043A\u043E\u0435 "\u043F\u0440\u0430\u0432\u0430"?
 memberPermissions.permissionHelp= 
 <b>\u0410\u0434\u043C\u0438\u043D\u0438\u0441\u0442\u0440\u0430\u0442\u043E\u0440</b> - \u043C\u043E\u0436\u0435\u0442 \u0432\u0435\u0441\u0442\u0438 \u0437\u0430\u043F\u0438\u0441\u0438, \u0443\u043F\u0440\u0430\u0432\u043B\u044F\u0442\u044C \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u043E\u043C \u0438 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F\u043C\u0438.
@@ -704,3 +699,4 @@ yourWebsites.youAreInvited=\u0412\u044B \u043F\u0440\u0438\u0433\u043B\u0430\u04
 
 yourWebsites.declined=\u0412\u044B \u043E\u0442\u043A\u043B\u043E\u043D\u0438\u043B\u0438 \u043F\u0440\u0438\u0433\u043B\u0430\u0448\u0435\u043D\u0438\u0435 , \u0447\u0442\u043E\u0431\u044B \u0440\u0430\u0431\u043E\u0442\u0430\u0442\u044C \u0432 \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0435 [{0}]
 yourWebsites.resigned=\u0412\u044B \u043E\u0442\u043F\u0438\u0441\u0430\u043D\u044B \u043E\u0442 \u0434\u043D\u0435\u0432\u043D\u0438\u043A\u0430 [{0}]
+inviteMember.prompt=\u041f\u0440\u0438\u0433\u043b\u0430\u0448\u0435\u043d\u0438\u0435 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f \u043f\u0440\u0438\u0441\u043e\u0435\u0434\u0438\u043d\u0438\u0442\u044c\u0441\u044f \u043a \u044d\u0442\u043e\u043c\u0443 \u0434\u043d\u0435\u0432\u043d\u0438\u043a\u0443, \u0432\u044b\u0431\u0440\u0430\u0432 \u0435\u0433\u043e  \u0438\u043b\u0438 \u043f\u043e email \u0430\u0434\u0440\u0435\u0441\u0443, \u0443\u043a\u0430\u0437\u0430\u043d\u043d\u043e\u043c\u0443 \u043d\u0438\u0436\u0435, \u0434\u0430\u0432 \u0440\u0430\u0437\u0440\u0435\u0448\u0435\u043d\u0438\u0435 \u0434\u043b\u044f \u043d\u0438\u0445 \u0438 \u0438 \u043a\u043b\u0438\u043a\u043d\u0443\u0432  \u043a\u043d\u043e\u043f\u043a\u0443 "\u041e\u043f\u0440\u0430\u0432\u0438\u0442\u044c"   "\u041f\u0440\u0438\u0433\u043b\u0430\u0441\u0438\u0442\u044c".
